Refactor even more utilities into separate files
