+++ clowder/utility/git/printing.py
@@ -7,7 +7,7 @@ import sys
 
 from termcolor import colored, cprint
 
-from clowder.utility.clowder_utilities import execute_command
+from clowder.utility.execute import execute_command
 from clowder.utility.git.git_repo import GitRepo
 import clowder.utility.formatting as fmt
